requirements.txt
selenium>=4.0.0
python-dotenv
